Added lighter card embeds to consulting page
