+++ components/blocks/youtubePlaylist.tsx
@@ -3,7 +3,6 @@
 import axios from "axios";
 import { Key, useEffect, useState } from "react";
 import { BsArrowRightCircle } from "react-icons/bs";
-import { Template, TinaField } from "tinacms";
 import { tinaField } from "tinacms/dist/react";
 import { VideoLink } from "../../services/server/youtube";
 import { CustomLink } from "../customLink";
@@ -79,63 +78,4 @@ export const YoutubePlaylistBlock: React.FC<YoutubePlaylistProps> = (props) => {
       )}
     </>
   );
-};
-
-export const youtubePlaylistSchema: TinaField = {
-  type: "object",
-  label: "Youtube Playlist",
-  name: "youtubePlaylist",
-  ui: {
-    itemProps: (item) => {
-      return { label: item?.title };
-    },
-  },
-  fields: [
-    {
-      type: "string",
-      label: "Title",
-      name: "title",
-    },
-    {
-      type: "string",
-      label: "Playlist Id",
-      name: "playlistId",
-      required: true,
-    },
-    {
-      type: "number",
-      label: "Videos Count",
-      name: "videosCount",
-      required: true,
-    },
-    {
-      type: "object",
-      label: "Playlist Button",
-      name: "playlistButton",
-      fields: [
-        {
-          type: "string",
-          name: "text",
-          label: "Text",
-        },
-        {
-          type: "string",
-          name: "link",
-          label: "Link",
-          description: "DEFAULT: PlaylistId.",
-        },
-        {
-          type: "boolean",
-          name: "animated",
-          label: "Animated?",
-        },
-      ],
-    },
-  ],
-};
-
-export const youtubePlayListBlockSchema: Template = {
-  name: "YoutubePlaylistBlock",
-  label: "Youtube PlayList Block",
-  fields: [youtubePlaylistSchema],
 };
