+++ tina/collections/usergroup.tsx
@@ -4,7 +4,7 @@ import { pageBlocks as sectionPageBlocks } from "../../components/usergroup/sect
 import { seoSchema } from "../../components/util/seo";
 
 import type { Collection } from "tinacms";
-import { youtubePlaylistSchema } from "../../components/blocks/youtubePlaylist";
+import { youtubePlaylistSchema } from "../../components/blocks/youtubePlaylist.schema";
 import { joinAsPresenterSchema } from "../../components/usergroup/joinAsPresenter";
 import { latestTechSchema } from "../../components/usergroup/latestTech";
 import azureBannerSchema from "../../components/util/showAzureBanner";
